requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.4.0
django-filter==24.3.0